+++ program-analysis/echidna/example/magic.sol
@@ -1,13 +1,15 @@
 contract C {
-  bool value_found = false;
-  function magic(uint magic_1, uint magic_2, uint magic_3, uint magic_4) public {
-    require(magic_1 == 42);
-    require(magic_2 == 129);
-    require(magic_3 == magic_4+333);
-    value_found = true;
-    return;
-  }
-  function echidna_magic_values() public returns (bool) {
-    return !value_found;
-  }
+    bool value_found = false;
+
+    function magic(uint256 magic_1, uint256 magic_2, uint256 magic_3, uint256 magic_4) public {
+        require(magic_1 == 42);
+        require(magic_2 == 129);
+        require(magic_3 == magic_4 + 333);
+        value_found = true;
+        return;
+    }
+
+    function echidna_magic_values() public returns (bool) {
+        return !value_found;
+    }
 }
