+++ program-analysis/manticore/exercises/exercise2/overflow.sol
@@ -1,8 +1,7 @@
-
 contract Overflow {
-    uint public sellerBalance=0;
+    uint256 public sellerBalance = 0;
 
-    function add(uint value) public returns (bool){
+    function add(uint256 value) public returns (bool) {
         sellerBalance += value; // complicated math, possible overflow
     }
 }
